Adding middleware for staff authentication/redirection
Adds middleware for all `/staff` routes in order to the check the
`req.session.user` variable for one of the staff roles. This makes it so that
when not signed in as a volunteer you cannot go directly to the path
`/staff/volunteer/account/{some id}` and edit some other volunteers information
without being a staff member. This allows us to take out the check in the route
in `server/router/routes/staffHomePage.js`.

We can also extend this to include finer tuned path and role checking for
different levels of staff.