Adding node-uuid module dependency to package.json.
This commit adds an additional node module dependency on the `node-uuid` module. The module is available on GitHub at:

    https://github.com/broofa/node-uuid

This module is used to generate version 1 or version 4 UUIDs as defined in RFC4122. It provides cryptographically strong random number generation on supporting platforms. So that we can create unique *and* random IDs. For example, this is useful to use for the team ids instead of MongoDB ObjectID objects. Because MongoDB ObjectIDs are roughly time-based and sequential, they are predictable. If we did use MongoDB ObjectIDs someone could take their own team id and use it to figure out team ids for other teams and join other teams without having been given the invite link for those teams.

The solution is to use version 4 UUIDs with the node-uuid module. You can also create time based IDs which are similar to the MongoDB ObjectIDs:

**Example**:

    // Generate a v1 (time-based) id
    uuid.v1(); // -> '6c84fb90-12c4-11e1-840d-7b25c5ee775a'

    // Generate a v4 (random) id
    uuid.v4(); // -> '110ec58a-a0f2-4ac4-8393-c866d813b8d1'

More Information on UUIDs:

    http://en.wikipedia.org/wiki/Universally_unique_identifier