Updating dbman module with new connection method
I update the dbman module to use a the `open` method of `MongoClient` instead of the `connect` method we were using before. This is different connection method where we instantiate the Server and authenticate against the database ourselves. You can get more information at:

http://mongodb.github.io/node-mongodb-native/driver-articles/mongoclient.html

The reason for doing this is that it may help us when we try to do such things as allowing the staff to clear the database. Establishing connections this way allows us to create the database if it does not exist and does not require us to have the database created before hand. This method will also allow us to establish connections to multiple databases which may be needed if we're going to allow them to clear the database from the front-end.