Fixing typos and controller bug in error callback
I realized that there was a bug happening because the `error` callback
to the `ajaxForm` function in the controllers actually returns the
`XMLHttpRequest` AJAX request object. This means in order to access the
string passed in to `res.send` you actually have to access the
responseText property of the returned `XMLHttpRequest` object.

Also fixed lots of typos in various other files.