Fixing bug in zipcodes route, null textarea
The zip code route was not functioning correctly in its state before
this commit. Regardless of whether or not the zip codes passed
validation, they were still attempted to be inserted into the array. The
route also attempts to send multiple responses for a single request in
certain circumstances. Also no need to have a nested loop to go over
each character in the zip code, a single test against a regular
expression for 5 digits simplifies things a lot. Also fixed the null
textarea problem in the view as well.