Added _layoutFile to app.locals call to remove it from all res.render calls
Added the entry `_layoutFile: 'default'` too the call to `app.locals` in the
file `app.js` so that it could be removed from all the calls to `res.render`
in each of the routes. This means that whenever `res.render` is called it will
always use `default` for `_layoutFile`. This is simpler and it will be more
efficient because `_layoutFile` does not have to be set for each
request/response. We have the design choice of whether or not to create multiple
layouts for different user types or to use logic within our templates in order
to change the single layout based on user type. If you want to use a different
layout then you can override `default` by passing in a new `_layoutFile` to
`res.render` as we've been doing.