Fixing bug in geocoder test taking changing data into account
The exact coordinates returned for a geocoding request for Auburn's address
has changed twice since the test was created. In order to fix this I am
changing to use the `closeTo` method in the chai assertion library.

The documentation that helped me implement this fix to make the geocoder tests
pass again can be found at:

http://chaijs.com/api/bdd/#closeTo-section