Fix for staff home page after database schema change.
This is a fix for left over previously undiscovered issues in the staff home page tables which result from the change in the database schema. The volunteer table was not sorting or loading rows properly. This fixes both of these problems for the `volunteer` table as well as the `cooordinator`, `committee`, and `leadership` tables.

As far as I can tell this will work for now, however, things are getting a little messy and it may be best to try to refactor this before making it anymore complex.

The problem was that most of the other fields in the volunteer table are now fields in the `volunteer` subdocument. The key that gets passed into the controllers `sort` function had to be changed depending on whether it was a user role type table. Also some additional logic was needed in the `replaceRow` function in the controller. I also changed `projectCoordinator` to be just `coordinator` in order to make things more consistent with roles.