Changes separate last and first name fields to be a single field
This commit illustrates how easy it is to change what fields the tables will
have. Each table has an array for its header labels, and each table has a
function which given a document, returns an array which is used as the row
in the table for that document. This is the only place changes need to be made
and decouples the implementation of the tables on the client side from the
content actually presented in the table.